[#9] Refactor: PayrollAction 컴포넌트 분리
