+++ src/pages/salary-management/salary-management-page.jsx
@@ -1,13 +1,13 @@
 import { Headding } from '@pages/home/home-page'
 import Payslip from './Payslip/Payslip'
-// import Payslip2 from './Payslip/Payslip2'
+import Payslip2 from './Payslip/Payslip2'
 
 function SalaryManagementPage() {
   return (
     <>
       <Headding>급여관리</Headding>
       <Payslip />
-      {/* <Payslip2 /> */}
+      <Payslip2 />
     </>
   )
 }
